User model changed to Dictionaries
The user is now just a dynamically built dictionary around the
categories that are specified in the main company XML, and their
association to any categories that are found in the searching of the
specified Excel Spreadsheet.

Next thing is to move test code into system.UserInfo so that we can
generate instant user lists from an Engine source file.

An additional method should be built to extract userlist from a
specified Excel File.